Restructure files and organization for better Components.
