+++ client/src/components/RelatedItems/RelatedProducts/ProductCard.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function ProductCard () {
+
+  return (
+    <div>
+      <h3>ProductCard</h3>
+    </div>
+  )
+}
+export default ProductCard;
